backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0